requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.7.1
Django==2.2
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.5
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
